Separate and extend numerics
